fix: decimal calories, FatSecret resilience, cockpit faith perf
- Nutrition: accept decimal calories (712.5) across form, AI intent,
  and action handler — DB already supports DecimalField
- FatSecret: add _safe_json() to handle empty API responses gracefully
  instead of crashing on JSONDecodeError
- Cockpit: replace 16 × get_execution_truth() calls (~560 queries)
  with 3-4 batch queries for faith completion dates

+++ apps/ai/action_handlers.py
@@ -918,7 +918,7 @@ class ActionHandler:
             )
 
     def handle_log_food(self, food_name: str, quantity: float = 1,
-                        meal_type: str = None, calories: int = None,
+                        meal_type: str = None, calories: float = None,
                         notes: str = "", **kwargs) -> ActionResult:
         """
         Log a food entry with smart nutrition lookup.
